src/test/java/com/example/tobyspringsix/payment/PaymentTest.java
package com.example.tobyspringsix.payment;

import static org.assertj.core.api.Assertions.*;

import java.math.BigDecimal;
import java.time.Clock;
import java.time.Duration;
import java.time.Instant;
import java.time.LocalDateTime;
import java.time.ZoneId;
import java.time.temporal.ChronoUnit;

import org.junit.jupiter.api.Test;

public class PaymentTest {

	@Test
	void createPrepared() {
		Clock clock = Clock.fixed(Instant.now(), ZoneId.systemDefault());

		Payment payment = Payment.createPrepared(1L, "USD", BigDecimal.TEN, BigDecimal.valueOf(1_000),
			LocalDateTime.now());

		assertThat(payment.getConvertedAmount()).isEqualByComparingTo(BigDecimal.valueOf(10_000));
		assertThat(payment.getValidUntil()).isEqualTo(LocalDateTime.now(clock).plusMinutes(30));
	}

	@Test
	void isValid() {
		Clock clock = Clock.fixed(Instant.now(), ZoneId.systemDefault());

		Payment payment = Payment.createPrepared(1L, "USD", BigDecimal.TEN, BigDecimal.valueOf(1_000),
			LocalDateTime.now(clock));

		assertThat(payment.isValid(clock)).isTrue();
		assertThat(payment.isValid(Clock.offset(clock, Duration.of(30, ChronoUnit.MINUTES)))).isFalse();
	}
}
